initialCommit
This is a simple Python program that uses the psutil library to display system information on Linux and Windows. The program displays the following information:

System temperature (Linux only)
Battery status (Windows only)
Last boot time
Number of running processes
Number of CPU cores and logical processors
Total and used memory
Disk partitions and their usage
Current users

The program also uses the clamd library to perform a quick virus scan on the system. Note that this requires the ClamAV antivirus software to be installed and running on the system.

python

import os
import psutil
import clamd

def get_system_info():
    uname = os.uname()
    if uname.sysname == 'Linux':
        temperature = psutil.sensors_temperatures()['cpu_thermal'][0].current
    elif uname.sysname == 'Windows':
        battery = psutil.sensors_battery()
        temperature = None
    else:
        raise ValueError("Unsupported operating system")

    boot_time = psutil.boot_time()
    boot_time_datetime = datetime.datetime.fromtimestamp(boot_time)

    processes = psutil.process_iter()
    process_count = len(list(processes))

    cpu_count = psutil.cpu_count()
    cpu_logical = psutil.cpu_count(logical=True)

    memory = psutil.virtual_memory()
    memory_total = memory.total / (1024 * 1024)
    memory_used = memory.used / (1024 * 1024)

    disk_partitions = psutil.disk_partitions()
    disk_usage = psutil.disk_usage('/')
    disk_used = disk_usage.used / (1024 * 1024)
    disk_total = disk_usage.total / (1024 * 1024)

    users = psutil.users()

    return {
        'system': uname.sysname,
        'temperature': temperature,
        'battery': battery.percent if battery else None,
        'boot_time': boot_time_datetime.strftime('%Y-%m-%d %H:%M:%S'),
        'process_count': process_count,
        'cpu_count': cpu_count,
        'cpu_logical': cpu_logical,
        'memory_total': memory_total,
        'memory_used': memory_used,
        'disk_used': disk_used,
        'disk_total': disk_total,
        'users': users
    }

def scan_system(clamd_host='localhost', clamd_port=3310):
    clamd_socket = clamd.ClamdUnixSocket(clamd_host, clamd_port)
    scan_result = clamd_socket.scan(' /')
    return scan_result

if __name__ == '__main__':
    system_info = get_system_info()
    scan_result = scan_system()

    print("System information:")
    print(json.dumps(system_info, indent=2))
    print("Virus scan result:", scan_result['result'])

Please Note that this is just a basic example, and there are many other system metrics that could be added. Also, it's important to handle exceptions and errors that may occur when querying system metrics. Finally, the virus scan is performed on the entire system, which can take a long time. It's recommended to scan specific directories or files instead.